requirements.txt
chainlit>=2.0.0
claude-code-sdk>=0.0.20
anthropic>=0.50.0
psycopg2-binary>=2.9.0
boto3>=1.35.0
python-dotenv>=1.0.0
